# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DetectRequest": {
        "examples": [
          {
            "image_path": "sample/higher-resolution/dark/14581.jpg",
            "plate_type": "petri",
            "polarity": "auto",
            "sensitivity": 50
          }
        ],
        "properties": {
          "adaptive_scale": {
            "default": false,
            "description": "1차 검출로 콜로니 크기를 재고 해상도를 자동 조정해 재검출한다. **검출 비용이 약 2배가 되고 실측 F1 은 오히려 떨어진다.** 해상도는 `work_size` 로 직접 지정할 것.",
            "title": "Adaptive Scale",
            "type": "boolean"
          },
          "annotate": {
            "default": "all",
            "description": "표시 이미지 모드. `all` = 검출 전체를 빨강으로, `pick` = 피킹 대상(`pickable`)만 초록으로. **응답 JSON 의 `colonies` 는 어느 쪽이든 전부 동일하게 반환된다** — 그리기에만 영향한다.",
            "enum": [
              "all",
              "pick"
            ],
            "title": "Annotate",
            "type": "string"
          },
          "candidate_source": {
            "default": "union",
            "description": "후보 생성 방식. `union` = LoG ∪ 다중레벨 이진화(기본, 재현율 우선) — 전 구간에서 LoG 단독보다 +2.9~3.5%p 위다. `threshold` = 이진화 단독 — **정밀도 93% 이상 구간에서는 이쪽이 낫다** (계수/CFU 용도). `log` = LoG 단독(구동작).",
            "enum": [
              "union",
              "log",
              "threshold"
            ],
            "title": "Candidate Source",
            "type": "string"
          },
          "color_boost": {
            "default": 0.0,
            "description": "찍은 색 축의 가중치. **0 = 끔(기본).** 권장값 0.5~0.6.\n\n켜면 `t = max(t_밝기, boost × t_색)` 이라 **기본 동작보다 나빠질 수 없다** — 색을 잘못 찍어도 최악이 지금과 같다.\n\n실측(목표색을 제대로 찍은 경우, 정밀도/재현율/F1):\n가장 못 잡는 `vague` 10장 — 0: 70.2%/54.9%/61.6 → **0.6: 70.6%/59.0%/64.3**\n대조군 `bright` 4장 — 0: 83.3%/75.6%/79.3 → 0.6: 82.7%/77.9%/80.2\n\n**교환이 아니다.** vague 에서 재현율 +4.1%p 를 얻는데 정밀도가 +0.4%p 로 같이 오른다. 흐린 접시에서 먼저 켜볼 것.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Color Boost",
            "type": "number"
          },
          "colour_credit": {
            "default": 1.0,
            "description": "색이 뚜렷한 후보의 감도 요구치를 이 배율까지 깎아준다. 1.0 = 끔(기본, 유지 권장). **현재 설정(candidate_source=\"union\")에서 켜면 전 그룹이 나빠진다** — 실측 2.0배에서 정밀도 82.20%→59.82%, 재현율 75.7%→79.2%, F1 78.80→68.14 (그룹별 F1: lower 83.2→67.3, bright 80.7→74.6, dark 83.7→70.0, vague 61.6→52.7). 정밀도를 22%p 깎고 재현율은 3.5%p 만 얻으므로 오퍼레이터 UI 에 노출하지 말 것.",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Colour Credit",
            "type": "number"
          },
          "edge_margin": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "피킹 안전 여백 (오퍼레이터용 0~100). 접시·웰 경계에서 이만큼 안쪽만 피킹 대상으로 인정한다. 실제 적용된 px 값은 응답 `applied_params.pick_edge_margin` 에 담긴다.",
            "title": "Edge Margin"
          },
          "exclude_nested": {
            "default": false,
            "description": "`parent_id` 가 붙은 검출을 `colonies` 에서 빼고 `count` 도 줄인다. **기본 끔** — 실측 39장에서 F1 이 0.23 내려간다(제외 대상 30개 중 16개가 정답). 같은 콜로니를 두 번 집는 것을 막고 싶을 때만 켠다. 켜면 남는 검출의 `parent_id` 는 모두 `null` 이고 `id` 는 1부터 다시 매겨진다. `pick_top_n` 은 제외 **전** 전체 집합에서 상위 N 을 고르므로, 함께 쓰면 `pickable` 이 N 보다 적을 수 있다.",
            "title": "Exclude Nested",
            "type": "boolean"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "입력 이미지 base64. `data:image/...;base64,` 접두사를 붙여도 된다. 브라우저 흐름에서 쓰는 **운영 계약**이다. `image` 와 `image_path` 중 하나는 반드시 있어야 한다.",
            "title": "Image"
          },
          "image_format": {
            "default": "jpeg",
            "description": "`return_image` 응답 이미지 형식. 4000px 급 원본을 PNG 로 담으면 응답이 수십 MB 가 된다.",
            "enum": [
              "jpeg",
              "png"
            ],
            "title": "Image Format",
            "type": "string"
          },
          "image_max_width": {
            "default": 1600,
            "description": "`return_image` 응답 이미지 최대 폭(px). 0 = 원본 크기. 좌표(`colonies`)는 축소와 무관하게 **항상 원본 픽셀 기준**이며, 축소된 이미지 위에 겹치려면 `annotated_image_scale` 을 곱해야 한다.",
            "maximum": 8000.0,
            "minimum": 0.0,
            "title": "Image Max Width",
            "type": "integer"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "서버 로컬 파일 경로. **개발·튜닝 편의용**이며 서버와 파일시스템을 공유할 때만 동작한다. 운영 배포에서는 쓰지 말 것. `image` 와 함께 오면 이쪽이 우선한다.",
            "title": "Image Path"
          },
          "image_quality": {
            "default": 85,
            "description": "`return_image` JPEG 품질 (`image_format=\"jpeg\"` 일 때만).",
            "maximum": 100.0,
            "minimum": 30.0,
            "title": "Image Quality",
            "type": "integer"
          },
          "marker": {
            "default": "square",
            "description": "서버가 그리는 모든 표시 이미지(`return_image` · `/detect/preview` · `save_annotated`)의 마커 모양. 콜로니가 원형이라 원을 그리면 윤곽선과 겹쳐 구분이 어렵다 — 직선 테두리가 한천 텍스처 위에서 훨씬 잘 보인다.",
            "enum": [
              "square",
              "circle"
            ],
            "title": "Marker",
            "type": "string"
          },
          "mask_walls": {
            "default": true,
            "description": "접시·웰 경계에서 안쪽만 피킹 대상으로 인정한다(테두리 반점에 핀을 찍지 않게). False = 경계 제한 없음. **검출 자체가 아니라 `pickable` 판정에만 영향한다.**",
            "title": "Mask Walls",
            "type": "boolean"
          },
          "max_diam_frac": {
            "default": 0.0,
            "description": "콜로니 크기 상한 — 콜로니 지름 ÷ **기준 길이**(`min_diam_frac` 와 같은 분모, `applied_params.plate_size_ref`). 0 = 제한 없음(기본). **켤 때 주의** — 실측 콜로니 지름이 기준 길이의 1.2~45% 로 매우 넓어 어떤 상한도 진짜 큰 콜로니를 버린다. 상한이 막으려는 대상(뭉친 덩어리·접시 테두리)은 이미 `watershed_split` 과 접시 반지름 수축이 담당하므로 오퍼레이터 UI 에는 하한만 노출하는 것을 권한다.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Diam Frac",
            "type": "number"
          },
          "min_circularity": {
            "default": 0.0,
            "description": "둘레 기반 원형도 4πA/P² 하한. **기본 0 = 끔.** 경계 거칠기에 극도로 민감해 합집합 후보의 이진화 성분(둘레가 거칠다)을 부당하게 버렸다 — 끄자 재현율 74.4 → 75.7%. 모양 판정은 면적 기반 `min_roundness` 가 담당한다.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Circularity",
            "type": "number"
          },
          "min_diam_frac": {
            "default": 0.0,
            "description": "콜로니 크기 하한 — 콜로니 지름 ÷ **기준 길이**. 0 = 제한 없음(기본). 비율이라 해상도·카메라와 무관하다. 실측 분포 1.2~45%, 중앙값 7%. 기준 길이는 `plate_type=\"petri\"` 에서 접시 지름이지만, `well8` 이거나 접시 검출이 실패하면 **이미지 짧은 변**으로 폴백한다 — 서버가 실제로 쓴 값이 응답 `applied_params.plate_size_ref`(원본 px) 에 담기므로 \"5% = 몇 px\" 은 그것으로 계산할 것.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Diam Frac",
            "type": "number"
          },
          "min_fill": {
            "default": 0.45,
            "description": "채움율 하한 — 윤곽 면적 ÷ bounding box 면적. 0.60 으로 올리면 정밀도 92~96% 구간에서 +1.4~2.1%p 이므로 계수(CFU) 용도에 쓸 만하다. 기본 운영점에서는 이득이 없다.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Fill",
            "type": "number"
          },
          "min_rel_sat": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 1.5,
            "description": "색 차이 요구 — |내부채도 − 주변채도| 하한. 0 = 끔. 인쇄 글씨·데브리는 이 값이 0 근처(−0.1~1.7)이고 콜로니는 7~57 이라 **오검출을 걷어내는 가장 강한 레버**다. 실측: 0 → 77.5%/76.2%, **1.5 → 82.2%/75.7%(기본)**, 3.0 → 84.0%/74.7%, 6.0 → 88.2%/69.2%, 12.0 → 93.0%/60.8% (정밀도/재현율). 생략하면 서버 기본값. 무채색 이미지에서는 서버가 색 축을 끄므로 이 값이 무시된다 — 응답 `applied_params.has_chroma` 로 확인할 것. 0 이면 `colour_credit` 도 함께 무력화된다 — 내부적으로 이 값을 분모로 나누기 때문이다.",
            "title": "Min Rel Sat"
          },
          "min_roundness": {
            "default": 0.55,
            "description": "면적 ÷ 최소외접원 면적 하한 — **주된 모양 판정**. 0 = 끔. **완화를 권하지 않는다.** 같은 정밀도(82.8%)에서 이 값을 0.45 로 푸는 것보다 감도를 내리는 쪽이 더 많이 맞힌다 (74.7% 대 72.5%).",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Roundness",
            "type": "number"
          },
          "min_solidity": {
            "default": 0.75,
            "description": "면적 ÷ convex hull 면적 하한. 오목한 얼룩·긁힘을 배제한다. 0 = 끔. 0.75 아래로는 결과가 변하지 않는다(포화).",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Solidity",
            "type": "number"
          },
          "min_t": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "감도 원본값 — 면적가중 Welch t-통계량 하한. 낮출수록 흐린 콜로니까지 잡아 재현율이 오르고 정밀도가 내린다. 실측: 30 → 89.0%/70.9%, 25 → 86.7%/73.3%, **20 → 82.2%/75.7%(기본)**, 15 → 73.5%/77.3% (정밀도/재현율). 생략하면 `sensitivity` 매핑, 그것도 없으면 서버 기본값(20)을 쓴다. **오퍼레이터 UI 는 이 값 대신 `sensitivity` 를 보낼 것.**",
            "title": "Min T"
          },
          "pick_radius_max": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "피킹 대상으로 인정할 최대 반지름(**원본 이미지 픽셀**). 생략하면 서버 기본값. 0 = 제한 없음.",
            "title": "Pick Radius Max"
          },
          "pick_radius_min": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "피킹 대상으로 인정할 최소 반지름(**원본 이미지 픽셀**). 생략하면 서버 기본값. 해상도가 다른 카메라로 바꾸면 재조정해야 한다.",
            "title": "Pick Radius Min"
          },
          "pick_top_n": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "피킹 후보 중 `score` 상위 N개만 `pickable=true` 로 남긴다 (예: 96핀 헤드 → 96). 생략하면 제한 없음. 1 이상. `colonies` 배열 자체는 줄지 않는다(`exclude_nested` 는 줄인다). 랭킹 기준인 `score` 는 **고립도와 원형도**로 결정된다 — `Colony.score` 설명 참조.",
            "title": "Pick Top N"
          },
          "plate_type": {
            "default": "petri",
            "description": "플레이트 기하구조. `petri` = 원형 접시(HoughCircles 로 ROI 검출), `well8` = 4×2 몰딩 8웰 플레이트(격자 ROI). 자동 판정은 판별력이 없어 호출자가 명시해야 한다.",
            "enum": [
              "petri",
              "well8"
            ],
            "title": "Plate Type",
            "type": "string"
          },
          "polarity": {
            "default": "auto",
            "description": "콜로니가 한천보다 밝은지 어두운지. `auto` = 접시별 자동 판정(기본, 실측 39장 판정 정확도 100%). `bright`/`dark` = 그 극성만 검출. `both` = 양극성 병합 — 자동 판정이 틀릴 때의 되돌림 경로. **기본값 유지를 권장한다.**",
            "enum": [
              "auto",
              "both",
              "bright",
              "dark"
            ],
            "title": "Polarity",
            "type": "string"
          },
          "return_image": {
            "default": false,
            "description": "표시 이미지를 `annotated_image`(base64)로 응답에 함께 담는다. **웹 UI 에서는 쓰지 말 것** — 4000px 이미지 기준 260KB 로 좌표만 받을 때(3.9KB)보다 67배 크고, 확대하면 마커가 뭉개진다. 좌표를 렌더링할 수 없는 클라이언트용이다.",
            "title": "Return Image",
            "type": "boolean"
          },
          "save_annotated": {
            "default": false,
            "description": "검출 표시 이미지를 서버의 `vision/output/` 에 저장하고 경로를 `annotated_path` 로 돌려준다. **개발·디버깅용** — 서버 디스크에 파일을 쓰므로 운영 UI 에서 쓰지 말 것.",
            "title": "Save Annotated",
            "type": "boolean"
          },
          "sensitivity": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "감도 (오퍼레이터용 0~100). 높을수록 흐린 콜로니까지 잡는다. `min_t` 로 매핑되며 `min_t` 를 직접 주면 그쪽이 우선한다. 실측: 43 → 89.0%/70.9%, 46 → 86.7%/73.3%, **50 → 82.2%/75.7%(기본)**, 81 → 73.5%/77.3% (정밀도/재현율). **매핑식을 클라이언트에 복제하지 말 것** — 실제 적용값이 응답 `applied_params.min_t` 로 되돌아온다.",
            "title": "Sensitivity"
          },
          "split_area_ratio": {
            "default": 1.5,
            "description": "윤곽 면적이 후보 기대 면적(π r²)의 이 배를 넘으면 병합으로 보고 분리한다. 낮출수록 적극적으로 나눈다. **반지름 보정과 함께 움직인다** — 반지름을 바꾸면 이 값도 재보정해야 한다.",
            "maximum": 5.0,
            "minimum": 0.5,
            "title": "Split Area Ratio",
            "type": "number"
          },
          "target_color": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "maxItems": 3,
                "minItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "오퍼레이터가 화면에서 찍은 콜로니 색 `[R, G, B]` (각 0~255). **이 색에 가까운 콜로니를 더 잘 찾는다** — 검출 후 거르는 게 아니라 검출 자체를 돕는다. `color_boost` 가 0 이면 무시된다.\n\n밝기를 뺀 색상 평면(Lab a\\*b\\*)에서 거리를 재므로 그늘져서 어두운 같은 색 콜로니도 인정한다.\n\n**색은 화면에서 찍게 할 것.** 숫자로 입력받으면 촬영마다 조명·화이트밸런스가 달라 같은 콜로니가 다른 값으로 찍힌다. 같은 이미지 안에서 고르면 그 문제가 없다.",
            "title": "Target Color"
          },
          "threshold_levels": {
            "default": 24,
            "description": "이진화 후보의 임계값 레벨 수 (`candidate_source` 가 union/threshold 일 때). 12/24/36 → 재현율 67.9/70.8/71.3% 로 **24 에서 포화한다** — 늘려도 커버리지 천장(약 90%)은 오르지 않고 비용만 는다.",
            "maximum": 64.0,
            "minimum": 2.0,
            "title": "Threshold Levels",
            "type": "integer"
          },
          "watershed_split": {
            "default": true,
            "description": "붙은 콜로니를 거리변환 watershed 로 분리한다. **끄면 나빠지기만 한다** — 이상하게 잘리는 접시가 있을 때만 쓰는 비상구이지 조정 knob 이 아니다.",
            "title": "Watershed Split",
            "type": "boolean"
          },
          "work_size": {
            "default": 1280,
            "description": "처리 해상도(최대변 px). 콜로니당 픽셀 수가 t-통계량을 좌우한다. 1280 이 실측 최적이고 2048 은 전 그룹에서 나쁘다. **바꾸면 크기 창 캘리브레이션이 무효가 된다** — 단독 변경 금지.",
            "maximum": 2048.0,
            "minimum": 384.0,
            "title": "Work Size",
            "type": "integer"
          }
        },
        "title": "DetectRequest",
        "type": "object"
      },
      "DetectResponse": {
        "properties": {
          "annotated_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "`return_image=true` 일 때 표시 이미지 base64.",
            "title": "Annotated Image"
          },
          "annotated_image_scale": {
            "default": 1.0,
            "description": "`annotated_image` 가 원본 대비 몇 배인지. 좌표를 그 이미지 위에 겹치려면 `x`/`y`/`radius` 에 이 값을 곱한다. 축소하지 않았으면 1.0.",
            "title": "Annotated Image Scale",
            "type": "number"
          },
          "annotated_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "`save_annotated=true` 일 때 서버에 저장된 이미지 경로.",
            "title": "Annotated Path"
          },
          "applied_params": {
            "default": {},
            "description": "이번 검출에 **서버가 실제로 적용한** raw 파라미터. `sensitivity` → `min_t` 처럼 매핑된 결과가 들어 있다. **UI 는 클라이언트에서 재계산하지 말고 이 값을 표시할 것.**",
            "title": "Applied Params",
            "type": "object"
          },
          "colonies": {
            "description": "검출된 콜로니 전체. 좌표는 항상 원본 픽셀 기준이므로 SVG `viewBox=\"0 0 {width} {height}\"` 에 그대로 얹으면 된다.",
            "items": {
              "$ref": "#/components/schemas/Colony"
            },
            "title": "Colonies",
            "type": "array"
          },
          "count": {
            "description": "검출된 콜로니 수 (`colonies` 길이와 같다).",
            "title": "Count",
            "type": "integer"
          },
          "height": {
            "description": "**원본 이미지** 높이(px). 처리 해상도와 무관하다.",
            "title": "Height",
            "type": "integer"
          },
          "width": {
            "description": "**원본 이미지** 폭(px). 처리 해상도와 무관하다.",
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "width",
          "height",
          "count",
          "colonies"
        ],
        "title": "DetectResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PreviewResponse": {
        "properties": {
          "count": {
            "description": "검출된 콜로니 수.",
            "title": "Count",
            "type": "integer"
          },
          "image": {
            "description": "검출 표시 이미지 base64 (PNG). 좌표는 담기지 않는다.",
            "title": "Image",
            "type": "string"
          }
        },
        "required": [
          "count",
          "image"
        ],
        "title": "PreviewResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "배양 플레이트 이미지에서 콜로니를 검출해 **원본 픽셀 좌표**를 반환한다.\n\n### 프론트엔드가 쓰는 두 엔드포인트\n\n| | |\n|---|---|\n| `GET /image?path=` | 배경으로 깔 원본 이미지 (브라우저가 캐시한다) |\n| `POST /detect` | 좌표 JSON — 4000px 이미지에서 약 3.9KB |\n\n좌표는 **처리 해상도와 무관하게 항상 원본 픽셀 기준**이므로\n`<svg viewBox=\"0 0 {width} {height}\">` 에 그대로 얹으면 좌표 변환이 필요 없다.\n\n`return_image: true` 로 서버가 그린 이미지를 받을 수도 있지만 응답이 260KB 로\n67배 커지고 확대하면 마커가 뭉개진다 — 웹 UI 에서는 쓰지 말 것.\n\n### 감도\n\n오퍼레이터가 실제로 조절하는 값은 `sensitivity`(0~100) 하나다. 나머지 기본값은\n`sample/` 라벨 39장(정답 1,886개) 실측 최적이다.\n\n| sensitivity | min_t | 정밀도 | 재현율 |\n|---|---|---|---|\n| 43 | 30 | 89.0% | 70.9% |\n| 46 | 25 | 86.7% | 73.3% |\n| **50 (기본)** | **20** | **82.2%** | **75.7%** |\n| 81 | 15 | 73.5% | 77.3% |\n\n**매핑식을 클라이언트에 복제하지 말 것.** 실제 적용된 값이 응답의\n`applied_params` 로 되돌아온다 — 복제 때문에 UI 표시가 서버와 어긋난 적이 네 번 있다.\n\n### 알아둘 것\n\n- 정밀도 82% · 재현율 76% 다. **검출 5~6개 중 1개는 빈 한천**이고 **4개 중 1개는\n  놓친다.** 로봇에 보내기 전 사람이 확인하는 단계를 두는 것을 전제로 한 값이다.\n- `image_path` 는 서버와 파일시스템을 공유할 때만 동작하는 **개발용**이다.\n  운영에서는 `image`(base64)를 쓴다.\n- 검출은 4000px 이미지에서 약 2초 걸린다. 감도 슬라이더는 디바운스할 것.\n",
    "title": "PICKABLE Vision Server",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/detect": {
      "post": {
        "description": "이미지에서 콜로니를 검출해 **원본 픽셀 좌표**를 반환한다.\n\n프론트엔드가 쓰는 주 엔드포인트다. 응답은 좌표만 담아 4000px 이미지에서도\n약 3.9KB 이고, 배경 이미지는 `GET /image` 로 따로 받아 브라우저가 캐시한다.\n\n좌표는 `work_size` 로 축소해 처리하더라도 **원본 픽셀로 되돌려 준다.**\n\n보낼 값은 보통 `sensitivity`(0~100) 하나다. 서버가 적용한 raw 값은\n`applied_params` 로 되돌아오므로 UI 는 그것을 표시하면 된다 —\n**매핑식을 클라이언트에서 재계산하지 말 것.**\n\n`pickable` 은 로봇이 안전하게 집을 수 있는 후보인지이고, `score` 는 그\n랭킹이다. 둘 다 **검출 신뢰도가 아니다.**",
        "operationId": "detect_colonies_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "이미지 디코딩 실패 또는 입력 누락"
          },
          "422": {
            "description": "파라미터 범위 위반"
          }
        },
        "summary": "콜로니 검출 → 좌표",
        "tags": [
          "detect"
        ]
      }
    },
    "/detect/preview": {
      "post": {
        "description": "검출 결과를 **그려 넣은 이미지**만 반환한다. 좌표는 담기지 않는다.\n\n파라미터를 눈으로 튜닝할 때 쓰는 개발용 엔드포인트다.\n**웹 UI 는 `POST /detect` + `GET /image` 를 쓸 것** — 그쪽이 67배 작고\n확대해도 마커가 뭉개지지 않는다.\n\n좌표와 이미지를 한 번에 받아야 하면 `POST /detect` 에\n`return_image: true` 를 주면 된다.",
        "operationId": "detect_preview_detect_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "이미지 디코딩 실패 또는 입력 누락"
          },
          "422": {
            "description": "파라미터 범위 위반"
          }
        },
        "summary": "콜로니 검출 → 표시 이미지 (좌표 없음)",
        "tags": [
          "detect"
        ]
      }
    },
...
    "/image": {
      "get": {
        "description": "UI 배경으로 깔 **원본 이미지**를 바이트 그대로 돌려준다.\n\n`POST /detect` 가 좌표만 반환하므로 클라이언트도 원본 이미지를 가져야\n오버레이를 그릴 수 있다. `<img src=\"/image?path=...\">` 로 쓰면 브라우저가\n캐시하므로 감도를 바꿔 재검출해도 이미지는 다시 받지 않는다.\n\n`path` 는 서버 실행 디렉터리 기준 상대 경로이고, 디렉터리 밖으로 나가거나\n이미지가 아닌 확장자면 403 이다.\n\n`/detect/preview` 와 다르다 — 그쪽은 검출 결과를 **그려 넣은** 이미지를\nbase64 로 주고, 이쪽은 표시 대상인 **원본**을 바이트로 준다.",
        "operationId": "serve_image_image_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/*": {}
            },
            "description": "이미지 바이트"
          },
          "400": {
            "description": "경로 형식 오류"
          },
          "403": {
            "description": "서버 루트 밖이거나 이미지 확장자가 아님"
          },
          "404": {
            "description": "파일 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "원본 이미지 바이트",
        "tags": [
          "image"
        ]
      }
    }
  },
...
}